rename environment parameter to unbreak hiera
This is a fix for #63.
We rename the `environment` to `venv_vars` in order to ensure that hiera
calls do *not* break, as soon as a puppet execution flow enters our
module. For consistency's sake, we change `environment` not only in the
main class (where it's definitely needed:
https://github.com/rodjek/puppet-lint/issues/574), but also in the
`certonly` define.