Cron should also run with space in certonly title
Issue #91 reports that creating a letsencrypt::certonly resource with a
title containing a space will create a cron entry that will never run.
That's because it will create a cron entry like this:

    10 22 * * * /cron_scripts_path/renew-title with space.sh

That executes /cron_scripts_path/renew-title that obviously does not
exist. When the full cron command is put in quotes, it will run as
expected.