Update OS support
Adds Debian 9 (Stretch), Debian 10 (Buster) and Fedora 30. Drops Debian
8 (Jessie) and Ubuntu 14.04.

Debian 8 is in LTS, but Debian backports doesn't support LTS
repositories. When Buster came out, the Jessie backports repository was
dropped. It can still install using VCS, but that's not supported by
this module and up to the user. Ubuntu 14.04 is EOL.