Add support for systemd service and timer units to renew
This also adds support for naming services to restart with systemd
instead of long-handing it all.

It also adds support for managing firewalls using firewalld for
systems that are not meant to be http/https accessible at all times.

These three are essentially the same support and it is difficult to
break it up.