Fix certbot VCS package installation
As referenced in #240 the official installation method was changed
in December 2020, deprecating the use of the `certbot-auto` script,
and recommending `snap` as the primary package source.

As `snap` isn't yet a builtin package provider in Puppet and I don't
fancy creating a dependency on a third party module, nor do I want
to add a load of `exec` commands for managing snap packages I chose
to fix the VCS install method, and since it's used in the CI there's
less to change.

Since we can no longer use the `certbot-auto` script we should
follow the setup instructions provided by Letsencrypt's development
guide https://certbot.eff.org/docs/contributing.html.  Their guide
simply states to call a python script which configures the virtual
environment, and since the script is gone we no longer need the
VENV_PATH environment variable.

As the command needs to create a python virtual environment we
need to add the required package to the VCS dependencies, and from
testing I found there was no need to call the initialisation command
when installing from a package source, so I've made it conditional
to the VCS installation.

I've also taken the opportunity to update the version of `certbot`
that's checked out during VCS installation since `0.39.0` is now
around 1.5 years old.  Version `1.7.0` is the latest version that
is supported by Python 3.5 which is the lowest Python3 version
that's installed by default via package managers.