Remove unneeded fqdn_rand_string() calls
According to the puppetlabs-stdlib documentation,
fqdn_rand_string()'s second argument is the character set to use
when generating the random string, and its third argument is
the seed.  The certonly class incorrectly passes the domain name
as the second argument, causing fqdn_rand_string() to use the
host's fqdn as the seed, resulting in more-frequent-than-necessary
hash collisions.  The current implementation of fqdn_rand() seems
to provide sufficient randomness, so fqdn_rand_string() doesn't
seem to be necessary.